Use Oracle extensions for named support
Rewrite `OracleNamedParameterJdbcTemplate` using named parameter support
in `OraclePreparedStatement`.

This has the advantage that all parsing code is gnone. As a disadvantage
all the collections support is now gone. Also null support is a bit wacky.
Also we only support `MapSqlParameterSource` and `BeanPropertySqlParameterSource`
but `SqlParameterSource` is so badly designed that this can't be helped.
We could in theory try to detect collections and in this case delegate to
super. Note that this would have the same `SqlParameterSource` limitations.