Add support for explicit statement caching
OJDBC has an API to explicitly cache certain prepared statements only.
This can be used to reduce the number of soft parses for frequently
used statements. Unfortunately the API is a bit cumbersome to use.
Creating the statement needs a lookup in the cache and closing the
statement needs to return it to the cache.
This can be mitigated with a custom PreparedStatementCreator that
performs the lookup and wraps the statement to intercept the close()
method.